Create the Code
Created the code: Import urllib, BeautifulSoup, and ssl. Prompt the User for url and use default if the user does not add a url (<1 character in length). Open and read the data, then run through BeautifulSoup, and retrieve all the anchor tags (links) and add them to a list.

Next, prompt the User to input the position number on the list they want and the number of times they want to repeat this process (the count). Loop through the data, adding the link in the correct position to the export_links list and decrease the count by one (to eventually get to zero). Set the url to be the newest identified link. 

Repeat the entire process with the new url. When loop reaches a count of 0 and ends, set ‘last_name’ to the last item (link) in the list. Find parameters using .find() to find the name within the link and apply them. Print the answer/the final name.